refactor(app): extract useShopFlow and useBattleControls (#316 phase 9a)
The two handler groups whose dependency sets are genuinely small.

- app/useShopFlow.ts — handleOpenPack (the 14-case quick-battle reward switch),
  handleBuyCrystalPack, handleCrystalsChanged, handlePackDone. The `packs`
  state and `packBackScreenRef` move inside: nothing outside this flow reads
  either, so the hook owns them and only returns `packs`.
- app/useBattleControls.ts — handlePlayCard, handlePlayAoeCard, handleSetStance,
  handleCycleSpeed, handleWaveRewardPick, handleWaveRewardSkip.

Refs App resets from every battle-start path (isTrainingModeRef, isCampaignRef,
campaignPlayCountsRef, battleUsedStructure, battleUsedMobileUnit) stay owned by
App and are passed in, same as phase 2 did for the telemetry refs.

App.tsx: 2,684 -> 2,543 lines. Build green, chunk boundaries unchanged,
2107 tests passing.
